src/requirements.txt
APScheduler==3.10.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.8
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzlocal==5.0.1
urllib3==2.0.5
Werkzeug==2.3.7
munch==3.0.0